Broke apart the sparta shared pointer allocator and shared pointer
